Refactored Accordion and Date into JS objects
I've been working in JavaScript for a while now, so I decided to apply the things that I've learned to update my Accordion widget code.
The widget is now built like a "traditional" JS object instead of as a bunch of global functions, and some common mistakes were also removed.